chore: split generated gql directories
