+++ src/cli/components/molecules/RedirectionStatusTable.tsx
@@ -1,7 +1,7 @@
 import { Box } from 'ink';
 import type { StoredResource } from 'kyrielle';
 
-import { FragmentType, unmask } from '../../../gql/index.js';
+import { FragmentType, unmask } from '../../gql/index.js';
 import { useStore$ } from '../../../utils/store.js';
 import RedirectionStatus, { RedirectionStatusItem } from '../atoms/RedirectionStatus.jsx';
