+++ src/client/janus-output.ts
@@ -4,8 +4,8 @@ import assert from 'node:assert';
 
 import { ConfigService } from '../config/config.service.js';
 import { Config, OutputConfig, RedirectionConfig } from '../config/type.js';
-import { graphql } from '../gql/index.js';
 import { generateRedirectionId } from '../utils/redirections.js';
+import { graphql } from './gql/index.js';
 import { JanusClient } from './janus-client.js';
 
 // Types
