+++ src/index.ts
@@ -1,1 +1,2 @@
 export { JanusClient, type JanusClientStatus } from './client/janus-client.js';
+export { JanusOutput } from './client/janus-output.js';
